Potential fix for pull request finding 'Result of multiplication cast to wider type'

+++ ta4j-examples/src/test/java/ta4jexamples/indicators/CachedIndicatorBenchmarkTest.java
@@ -259,7 +259,7 @@ public class CachedIndicatorBenchmarkTest {
                 checksum += future.join();
             }
 
-            long operations = threads * readsPerThread;
+            long operations = ((long) threads) * readsPerThread;
             return new ScenarioResult(operations, durationNanos, checksum);
         } finally {
             pool.shutdown();
